+++ DOCs/TEMPLATEs/model_driver_P_Template.c
@@ -124,7 +124,7 @@ static void compute(void* km, int* ier)
    int NBC;
    char* NBCstr;
 
-   intptr_t* nAtoms;
+   int* nAtoms;
    int* atomTypes;
    double* cutoff;
    double* cutsq;
@@ -238,7 +238,7 @@ static void compute(void* km, int* ier)
 
 
    /* unpack data from KIM object */
-   nAtoms = (intptr_t*) KIM_API_get_data(pkim, "numberOfAtoms", ier);
+   nAtoms = (int*) KIM_API_get_data(pkim, "numberOfAtoms", ier);
    if (1 > *ier)
    {
       report_error(__LINE__, "KIM_API_get_data", *ier);
