+++ EXAMPLE_LEGOS/model_legos/example_model_Ne_P_MLJ_NEIGH_RVEC_F/model_Ne_P_MLJ_NEIGH_RVEC_F.c
@@ -71,7 +71,7 @@ static void compute(void* km, int* ier)
    int comp_energyPerAtom;
    int comp_virial;
 
-   intptr_t* nAtoms;
+   int* nAtoms;
    int* nAtomTypes;
    int* atomTypes;
    double* cutoff;
@@ -92,7 +92,7 @@ static void compute(void* km, int* ier)
    int* neighListOfCurrentAtom;
 
    /* unpack data from KIM object */
-   nAtoms = (intptr_t*) KIM_API_get_data(pkim, "numberOfAtoms", ier);
+   nAtoms = (int*) KIM_API_get_data(pkim, "numberOfAtoms", ier);
    if (1 > *ier)
    {
       report_error(__LINE__, "KIM_API_get_data", *ier);
